More updates to Makefile, fixed float benchmarks to work in c

+++ benchmarks/float/addition.c
@@ -7,7 +7,7 @@ int main () {
     float adder = 0.0;
 
     for(int i =0; i < 10000; i++){
-        adder += static_cast <float> (rand() / (static_cast <float> (RAND_MAX / 1000.0)));
+        adder += (rand() / ((RAND_MAX / 1000.0)));
         adder = 0.0;
     }
