+++ benchmarks/float/division.c
@@ -7,7 +7,7 @@ int main () {
     float divider = 1000.0;
 
     for(int i =0; i < 10000; i++){
-        divider /= static_cast <float> (rand() / (static_cast <float> (RAND_MAX / 1000.0))) + 1.0;
+        divider /= (rand() / ( (RAND_MAX / 1000.0))) + 1.0;
         divider = 1000.0;
     }
